Fix a couple of conversion warnings in OptionParser for testing.

See merge request !31

+++ vtkm/testing/OptionParser.h
@@ -540,7 +540,7 @@ public:
    */
   int index() const
   {
-    return desc == 0 ? -1 : desc->index;
+    return desc == 0 ? -1 : static_cast<int>(desc->index);
   }
 
   /**
@@ -1454,7 +1454,7 @@ public:
       { return false; } // overflow protection: don't accept number of options that doesn't fit signed int
 
       buffer[parser.op_count] = option;
-      int idx = buffer[parser.op_count].desc->index;
+      unsigned int idx = buffer[parser.op_count].desc->index;
       if (options[idx])
       { options[idx].append(buffer[parser.op_count]); }
       else
